requirements.txt
pytest==8.2.2
Faker==25.9.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
selenium==4.21.0